[dev] outer proxy - cloudflare

+++ shared/scripts/services/cloudflare.sh
@@ -36,7 +36,11 @@ while getopts ':-:' OPT; do
 		task_info ) arg_task_info="${OPTARG:-}";;
 		toolbox_service ) arg_toolbox_service="${OPTARG:-}";;
 		webservice ) arg_webservice="${OPTARG:-}";;
-		output_file ) arg_output_file="${OPTARG:-}";;
+		output_file_relpath ) arg_output_file_relpath="${OPTARG:-}";;
+		only_if_needed )
+			arg_only_if_needed="${OPTARG:-}";
+			[ -z "${OPTARG:-}" ] && arg_only_if_needed='true'
+			;;
 		??* ) error "$command: Illegal option --$OPT" ;;  # bad long option
 		\? )  exit 2 ;;  # bad short option (error reported via getopts)
 	esac
@@ -49,46 +53,55 @@ title="${title}${command}"
 
 case "$command" in
 	"service:cloudflare:ips")
+		output_file_outer="/var/main/data/${arg_output_file_relpath:-}"
+		output_file_inner="/var/main/data/${arg_output_file_relpath:-}"
+
 		tmp_dir_inner='/tmp/main/cloudflare'
 		reload_file_inner="$tmp_dir_inner/reload"
 		reload_file_outer="$pod_data_dir/tmp/cloudflare/reload"
 
-		if [ -z "${arg_webservice:-}" ]; then
-			error "$command: webservice not defined"
+		if [ -z "${arg_output_file_relpath:-}" ]; then
+			error "$command: output_file_relpath not defined"
 		fi
 
-		"$pod_script_env_file" exec-nontty "$arg_toolbox_service" /bin/bash <<-SHELL || error "$command"
-			set -eou pipefail
-
-			mkdir -p "$tmp_dir_inner"
-			wget https://www.cloudflare.com/ips-v4 -O "$tmp_dir_inner"/ips-v4.txt >&2
-			wget https://www.cloudflare.com/ips-v6 -O "$tmp_dir_inner"/ips-v6.txt >&2
-			cat "$tmp_dir_inner"/ips-v4.txt "$tmp_dir_inner"/ips-v6.txt > "$tmp_dir_inner"/ips.txt
-
-			echo "ip rules..." >&2
-			sed 's/$/ 1;/' "$tmp_dir_inner"/ips.txt > "$tmp_dir_inner"/ips-rules.txt
-
-			touch "$arg_output_file"
-
-			checksum1="$(md5sum "$arg_output_file" | awk '{print $1}')"
-			checksum2="$(md5sum "$tmp_dir_inner"/ips-rules.txt | awk '{print $1}')"
-
-			if [ "\$checksum1" = "\$checksum2" ]; then
-				echo "rules unchanged - skipping..." >&2
-			else
-				echo "rules changed - generate final file..." >&2
-				mkdir -p "$(basedir "$arg_output_file")"
-				cp "$tmp_dir_inner"/ips-rules.txt "$arg_output_file"
-				touch "$reload_file_inner"
-				chmod 666 "$reload_file_inner"
-			fi
-		SHELL
-
-		if [ -f "$reload_file_outer" ]; then
-			echo "reloading $arg_webservice..." >&2
-			"$pod_script_env_file" "service:${arg_webservice}:reload"
+		if [ ! -f "$output_file_outer" ] || [ "${arg_only_if_needed:-}" != 'true' ]; then
 			"$pod_script_env_file" exec-nontty "$arg_toolbox_service" \
-				rm "$reload_file_inner"
+					/bin/bash <<-SHELL || error "$command"
+				set -eou pipefail
+
+				mkdir -p "$tmp_dir_inner"
+				wget https://www.cloudflare.com/ips-v4 -O "$tmp_dir_inner"/ips-v4.txt >&2
+				wget https://www.cloudflare.com/ips-v6 -O "$tmp_dir_inner"/ips-v6.txt >&2
+				cat "$tmp_dir_inner"/ips-v4.txt "$tmp_dir_inner"/ips-v6.txt > "$tmp_dir_inner"/ips.txt
+
+				echo "ip rules..." >&2
+				sed 's/$/ 1;/' "$tmp_dir_inner"/ips.txt > "$tmp_dir_inner"/ips-rules.txt
+
+				touch "$output_file_inner"
+
+				checksum1="\$(md5sum "$output_file_inner" | awk '{print \$1}')"
+				checksum2="\$(md5sum "$tmp_dir_inner"/ips-rules.txt | awk '{print \$1}')"
+
+				if [ "\$checksum1" = "\$checksum2" ]; then
+					echo "rules unchanged - skipping..." >&2
+				else
+					echo "rules changed - generate final file..." >&2
+					mkdir -p "$(dirname "$output_file_inner")"
+					cp "$tmp_dir_inner"/ips-rules.txt "$output_file_inner"
+
+					if [ -n "${arg_webservice:-}" ]; then
+						touch "$reload_file_inner"
+						chmod 666 "$reload_file_inner"
+					fi
+				fi
+			SHELL
+
+			if [ -f "$reload_file_outer" ] && [ -n "${arg_webservice:-}" ]; then
+				echo "reloading $arg_webservice..." >&2
+				"$pod_script_env_file" "service:${arg_webservice:-}:reload"
+				"$pod_script_env_file" exec-nontty "$arg_toolbox_service" \
+					rm "$reload_file_inner"
+			fi
 		fi
 		;;
 	*)
